requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
supabase==2.4.3
python-dotenv==1.0.0
